google/requirements.txt
astroid @ file:///C:/ci/astroid_1592481955828/work
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
et-xmlfile==1.0.1
googlemaps==4.4.1
idna==2.9
install==1.3.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
pycodestyle==2.6.0
pylint @ file:///C:/ci/pylint_1592487534522/work
python-dateutil==2.8.1
pytz==2021.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
typed-ast==1.4.1
urllib3==1.25.9
wincertstore==0.2
wrapt==1.11.2
xlrd==2.0.1
XlsxWriter==1.2.9
python-dotenv==0.19.0
